Add util file
