+++ headers/util.h
@@ -1,0 +1,6 @@
+#ifndef UTIL_H
+#define UTIL_H
+
+void print_prompt();
+
+#endif
